Move ContainerDef and polish code
